requirements.txt
torch==2.10.0
numpy==2.4.2
pandas==3.0.1
polars==1.38.1
matplotlib==3.10.8
scikit_learn==1.8.0
tqdm==4.67.3
PyYAML==6.0.3